pythonGUI/requirements.txt
torch~=2.6.0
PyQt5~=5.15.11
torchvision~=0.21.0
pillow~=11.2.1
kagglehub~=0.3.11
pandas~=2.2.3
matplotlib~=3.9.4
seaborn~=0.13.2
scikit-learn~=1.6.1